feat(swagger): restructure swagger-related imports and create modular files for better organization and maintainability feat(swagger): implement swaggerApiDocument, swaggerApiSave, and swaggerApiSetup functions to enhance API documentation capabilities
